put logic into the factory
