+++ src/Pim/Component/Catalog/TurnProductIntoVariantProduct.php
@@ -23,13 +23,8 @@ class TurnProductIntoVariantProduct
             throw new \Exception('Product and product model families should be the same.');
         }
 
-        $variantProduct = VariantProduct::fromProduct($product);
+        $variantProduct = $this->createVariantProduct($product);
 
-        return $variantProduct;
-    }
-
-    public function filter(ProductInterface $product, ProductModelInterface $parent)
-    {
         $parentValues = $parent->getValues();
         $filteredValues = $product->getValues()->filter(
             function (ValueInterface $value) use ($parentValues) {
@@ -37,8 +32,38 @@ class TurnProductIntoVariantProduct
             }
         );
 
-        $product->setValues($filteredValues);
+        $variantProduct->setValues($filteredValues);
 
-        return $product;
+        return $variantProduct;
+    }
+
+    /**
+     * @param ProductInterface $product
+     *
+     * @return VariantProductInterface
+     */
+    private function createVariantProduct(ProductInterface $product)
+    {
+        //TODO: inject the class
+        $variantProduct = new VariantProduct();
+
+        $valueIdentifier = $product->getValues()->filter(
+            function (ValueInterface $value) {
+                return AttributeTypes::IDENTIFIER === $value->getAttribute()->getType();
+            }
+        )->first();
+
+        $variantProduct->setId($product->getId());
+        $variantProduct->setIdentifier($valueIdentifier);
+        $variantProduct->setGroups($product->getGroups());
+        $variantProduct->setAssociations($product->getAssociations());
+        $variantProduct->setEnabled($product->isEnabled());
+        $variantProduct->setCompletenesses($product->getCompletenesses());
+        $variantProduct->setFamily($product->getFamily());
+        $variantProduct->setCategories($product->getCategories());
+        $variantProduct->setCreated($product->getCreated());
+        $variantProduct->setUpdated($product->getUpdated());
+
+        return $variantProduct;
     }
 }
