+++ src/Pim/Component/Catalog/spec/Model/VariantProductSpec.php
@@ -219,67 +219,4 @@ class VariantProductSpec extends ObjectBehavior
             'otherValue-<all_channels>-<all_locales>' => $otherValue
         ]);
     }
-
-    public function it_creates_a_variant_product_from_a_product(
-        ProductInterface $product,
-        Collection $groups,
-        Collection $associations,
-        Collection $completenesses,
-        Collection $categories,
-        FamilyInterface $family,
-        \Datetime $createdAt,
-        \Datetime $updatedAt,
-        ValueCollectionInterface $values,
-        ValueCollectionInterface $skuValues,
-        ValueInterface $valueDescription,
-        ValueInterface $valueSku,
-        AttributeInterface $description,
-        AttributeInterface $sku
-    ) {
-        $categories->toArray()->willReturn([]);
-        $values->toArray()->willReturn([]);
-
-        $description->getType()->willReturn(AttributeTypes::TEXTAREA);
-        $description->getCode()->willReturn('description');
-        $description->isUnique()->willReturn(false);
-        $sku->getType()->willReturn(AttributeTypes::IDENTIFIER);
-        $sku->getCode()->willReturn('sku');
-        $sku->isUnique()->willReturn(true);
-
-        $valueSku->getData()->willReturn('foo');
-        $valueSku->getAttribute()->willReturn($sku);
-        $valueSku->getScope()->willReturn(null);
-        $valueSku->getLocale()->willReturn(null);
-        $valueDescription->getAttribute()->willReturn($description);
-        $valueDescription->getScope()->willReturn(null);
-        $valueDescription->getLocale()->willReturn(null);
-
-        $values->filter(Argument::any())->willReturn($skuValues);
-        $skuValues->first()->willReturn($valueSku);
-
-        $product->getId()->willReturn(42);
-        $product->getGroups()->willReturn($groups);
-        $product->getAssociations()->willReturn($associations);
-        $product->isEnabled()->willReturn(true);
-        $product->getCompletenesses()->willReturn($completenesses);
-        $product->getFamily()->willReturn($family);
-        $product->getCategories()->willReturn($categories);
-        $product->getValues()->willReturn($values);
-        $product->getCreated()->willReturn($createdAt);
-        $product->getUpdated()->willReturn($updatedAt);
-
-        $result = $this::fromProduct($product);
-        $result->shouldReturnAnInstanceOf(VariantProductInterface::class);
-        $result->getId()->shouldReturn(42);
-        $result->getIdentifier()->shouldReturn('foo');
-        $result->getGroups()->shouldReturn($groups);
-        $result->getAssociations()->shouldReturn($associations);
-        $result->isEnabled()->shouldReturn(true);
-        $result->getCompletenesses()->shouldReturn($completenesses);
-        $result->getFamily()->shouldReturn($family);
-        $result->getCategories()->shouldReturnAnInstanceOf(Collection::class);
-        $result->getValues()->shouldReturnAnInstanceOf(ValueCollectionInterface::class);
-        $result->getCreated()->shouldReturn($createdAt);
-        $result->getUpdated()->shouldReturn($updatedAt);
-    }
 }
